Major fixes: DB seeding via raw SQL, editable Settings, 8.5h work hours, dynamic exit distance
- DatabaseModule: Fix DB seeding using raw SQL instead of 2nd Room instance
- Settings: Add editable sliders for work hours, geofence radius, exit distance
- Work hours default changed to 8.5 (8h 30m)
- TriggerManager: exit_watch_distance from VariableStore instead of hardcoded
- VariableStore: getWorkDurationHours returns Float for fractional hours
- SettingsViewModel: load/save from SharedPreferences + sync to VariableStore

+++ app/src/main/java/com/automate/engine/TaskRunner.kt
@@ -358,11 +358,11 @@ class TaskRunner @Inject constructor(
 
     // === Smart Time-Out Flow ===
 
-    private fun scheduleTimeOutPrompt(workHours: Int) {
+    private fun scheduleTimeOutPrompt(workHours: Float) {
         timeOutJob?.cancel()
         timeOutJob = scope.launch {
             // Wait for work hours to complete
-            delay(workHours * 60 * 60 * 1000L)
+            delay((workHours * 60 * 60 * 1000).toLong())
 
             Log.i(TAG, "Work hours complete, prompting for time-out")
             sendTimeOutPrompt()
